sidecar/app/requirements.txt
fastapi
uvicorn[standard]

opentelemetry-instrumentation
opentelemetry-instrumentation-logging
opentelemetry-distro[otlp,fastapi]
